+++ src/dolphin/unwrap/_constants.py
@@ -1,0 +1,2 @@
+CONNCOMP_SUFFIX = ".unw.conncomp.tif"
+UNW_SUFFIX = ".unw.tif"
